Change data type from initial grid from default (float) to intc

+++ gridworld.py
@@ -58,7 +58,8 @@ def init_grid():
 
 # Initialize player in random location, but keep wall, goal and pit stationary
 def init_grid_player():
-    state = np.zeros((4, 4, 5))
+    # TODO Do I want to use int as a type, so far I used the default
+    state = np.zeros((4, 4, 5), dtype=np.intc)
     # place player a
     state[randPair(0, 4)] = player_a.copy()
     # place player b
